+++ packages/common/src/chains/types.ts
@@ -1,5 +1,8 @@
 import type { Chain } from "viem/chains";
 
 export type MUDChain = Chain & {
+  iconUrls?: readonly string[];
+  indexerUrl?: string;
+  /** @deprecated */
   faucetUrl?: string;
 };