+++ packages/common/src/index.ts
@@ -7,6 +7,7 @@ export * from "./getBurnerPrivateKey";
 export * from "./getNonceManager";
 export * from "./getNonceManagerId";
 export * from "./hexToResource";
+export * from "./LruMap";
 export * from "./readHex";
 export * from "./resourceToLabel";
 export * from "./resourceToHex";