+++ packages/world/ts/config/defaults.ts
@@ -1,7 +1,7 @@
 export const SYSTEM_DEFAULTS = {
   registerFunctionSelector: true,
   openAccess: true,
-  accessList: [] as string[],
+  accessList: [],
 } as const;
 
 export type SYSTEM_DEFAULTS = typeof SYSTEM_DEFAULTS;
@@ -9,14 +9,13 @@ export type SYSTEM_DEFAULTS = typeof SYSTEM_DEFAULTS;
 export const WORLD_DEFAULTS = {
   worldContractName: undefined,
   worldInterfaceName: "IWorld",
-  systems: {} as Record<string, never>,
-  excludeSystems: [] as string[],
+  systems: {},
+  excludeSystems: [],
   postDeployScript: "PostDeploy",
   deploysDirectory: "./deploys",
   worldsFile: "./worlds.json",
   worldgenDirectory: "world",
   worldImportPath: "@latticexyz/world/src/",
-  modules: [] as [],
 } as const;
 
 export type WORLD_DEFAULTS = typeof WORLD_DEFAULTS;