+++ templates/react/packages/client/src/mud/supportedChains.ts
@@ -8,13 +8,12 @@
  *   basefee set to zero to avoid transaction fees.
  * - latticeTestnet, our public test network.
  *
-
  */
 
-import { MUDChain, latticeTestnet, mudFoundry } from "@latticexyz/common/chains";
+import { MUDChain, mudFoundry, redstone, garnet } from "@latticexyz/common/chains";
 
 /*
  * See https://mud.dev/tutorials/minimal/deploy#run-the-user-interface
  * for instructions on how to add networks.
  */
-export const supportedChains: MUDChain[] = [mudFoundry, latticeTestnet];
+export const supportedChains: MUDChain[] = [mudFoundry, redstone, garnet];