+++ packages/common/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 2.0.11
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.0.11
+
+## 2.0.10
+
+### Patch Changes
+
+- 51b137d3: Added OP predeploy contracts for Redstone and Garnet chain configs and added chain-specific contracts for Redstone chain config.
+  - @latticexyz/schema-type@2.0.10
+
+## 2.0.9
+
+### Patch Changes
+
+- 764ca0a0: Added an optional `indexerUrl` property to `MUDChain`, and populated it in the Redstone and Garnet chain configs.
+- bad3ad1b: Added chain icons to Redstone and Garnet chain configs via `chain.iconUrls`.
+  - @latticexyz/schema-type@2.0.9
+
 ## 2.0.8
 
 ### Patch Changes